Edit ReviewsAdapter.kt Edit Review.kt

+++ app/src/main/java/models/Review.kt
@@ -4,5 +4,5 @@ data class Review(
     val ID: Int,
     var author: String,
     var comment: String,
-    var ratingNum: Int,
+    var ratingNum: Float,
 )
